Progress on refactoring
