+++ src/scripts.js
@@ -7,6 +7,8 @@ import Activity from './Activity';
 import Sleep from './Sleep';
 import Chart from 'chart.js/auto';
 import {userData, userSleepData, userActivityData, userHydrationData} from './fetch.js';
+import {getLatestDate, getSleepComparison, getActivityComparisonData, getWeeklyAvgActivityData} from './utils.js';
+// import {generateStepGoalChart, generateWeekWaterChart, generateDayWaterChart, generateWeekSleepChart, generateAvgSleepChart, generateActivityComparisonChart} from './charts.js';
 import Hydration from './Hydration';
 
 const header = document.querySelector('#header')
@@ -16,7 +18,6 @@ const sleepChartAvg = document.querySelector('#sleepChartAvg')
 const waterChartWeek = document.querySelector('#waterChartWeek')
 const waterChartDay = document.querySelector('#waterChartDay')
 const activityComparisonChart = document.querySelector('#userAvgActivityComparisonChart')
-
 
 const fetchData = () => {
   return Promise.all([userData(), userSleepData(), userActivityData(), userHydrationData()])
@@ -31,32 +32,8 @@ const parseData = (data) => {
   loadPage([usersData, sleepEntries, activityData, hydrationData])
 }
 
-const getLatestDate = (dataset, user) => {
-  return dataset.reduce((latestDate, entry) => {
-    if (entry.userID === user.id && entry.date > latestDate) {
-      latestDate = entry.date;
-    }
-    return latestDate;
-  }, '')
-}
-
 const generateRandomIndex = (dataset) => {
   return Math.floor(Math.random() * dataset.length);
-}
-
-const getSleepComparison = (currentUser, sleepData, date) => {
-  const hours = currentUser.findHoursSleptByWeek(sleepData, date)[6].hoursSlept;
-  const quality = currentUser.findHoursSleptByWeek(sleepData, date)[6].sleepQuality;
-  const avgHours = currentUser.calculateAvgDailySleep(sleepData);
-  const avgQuality = currentUser.calculateAvgSleepQuality(sleepData);
-  const comparison =  {
-    date,
-    hoursSleptOnDate: hours,
-    sleepQualityOnDate: quality,
-    hoursSleptAvg: avgHours,
-    sleepQualityAvg: avgQuality
-  }
-  return comparison;
 }
 
 const generateHeaderContent = (user, stepsByDate, milesWalked, minutesActive, date, weeklyData) => {
@@ -168,7 +145,6 @@ const generateWeekWaterChart = (ouncesByWeek) => {
     }
   })
 }
-
 
 const generateDayWaterChart = (ouncesByDay) => {
   return new Chart(waterChartDay, {
@@ -323,26 +299,6 @@ const generateActivityComparisonChart = (comp) => {
       }
     }
   })
-}
-
-const getActivityComparisonData = (user, userRepo, activity, date) => {
-  return   {
-    userNumSteps: user.findStepsByDate(activity, date),
-    avgNumSteps: userRepo.calculateAvgStepsTaken(activity, date),
-    userMinActive: user.findMinsActiveByDate(activity, date),
-    avgMinActive: userRepo.calculateAvgMinActive(activity, date),
-    userFlights: user.findFlightsByDate(activity, date),
-    avgFlights: userRepo.calculateAvgStairsClimbed(activity, date),
-  }
-}
-
-const getWeeklyAvgActivityData = (user, activity, date) => {
-  return {
-    miles: user.calculateWeeklyMiles(activity, date),
-    numSteps: user.calculateWeeklySteps(activity, date),
-    minActive: user.calculateWeeklyActive(activity, date),
-    flights: user.calculateWeeklyFlights(activity, date),
-  }
 }
 
 const loadPage = (data) => {
